backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pillow==10.4.0
torch>=2.2.0
diffusers==0.30.3
transformers==4.44.2
accelerate==0.34.2
safetensors==0.4.5
